Fix DynamicClient.server_side_apply
DynamicClient.server_side_apply is designed to accept a dict or a
ResourceInstance as body.  However, if a dict or a ResourceInstance is
passed actually, an error occurs because RESTClientObject.rest cannot
interpret the Content-Type application/apply-patch+yaml.

So, modify RESTClientObject.rest to treat application/apply-patch+yaml
as other json-based Content-Types.